incorporated changes by Matthias
