+++ openml/flows/__init__.py
@@ -1,7 +1,8 @@
 from .flow import OpenMLFlow, _copy_server_fields
 
-from .sklearn_converter import sklearn_to_flow, flow_to_sklearn, _check_n_jobs
+from .sklearn_converter import sklearn_to_flow, flow_to_sklearn, \
+    _check_n_jobs, openml_param_name_to_sklearn
 from .functions import get_flow, list_flows, flow_exists, assert_flows_equal
 
 __all__ = ['OpenMLFlow', 'get_flow', 'list_flows', 'sklearn_to_flow',
-           'flow_to_sklearn', 'flow_exists', '_check_n_jobs']
+           'flow_to_sklearn', 'flow_exists', 'openml_param_name_to_sklearn']
