+++ openml/setups/__init__.py
@@ -1,6 +1,5 @@
 from .setup import OpenMLSetup, OpenMLParameter
 from .functions import get_setup, list_setups, setup_exists, initialize_model
-from .sklearn_converter import openml_param_name_to_sklearn
 
 __all__ = ['OpenMLSetup', 'OpenMLParameter', 'get_setup', 'list_setups',
-           'setup_exists', 'initialize_model', 'openml_param_name_to_sklearn']
+           'setup_exists', 'initialize_model']
